Commiting my changes so far for the integration tests

+++ tests/integration_tests/conftest.py
@@ -1,11 +1,13 @@
 import pytest
 import distutils
 import subprocess
-
+import os
+import atexit
+from filelock import FileLock
 from tests.utils import get_config_file_path
 
-@pytest.fixture(scope="class", autouse=True)
-def test_setup(request):
+@pytest.fixture(scope="session", autouse=True)
+def test_setup(tmp_path_factory, worker_id):
     def _test_setup():
         if distutils.spawn.find_executable("docker") and distutils.spawn.find_executable("terraform"):
             print("Setting up docker/terraform")
@@ -17,10 +19,27 @@ def test_setup(request):
                 print("Failed to setup docker/terraform for test, you must have Vault installed locally. Error: " + str(e))
 
     def _test_teardown():
-        print("Tearing down docker/terraform")
-        docker_file = get_config_file_path("vault-ldap/docker-compose.yml")
-        subprocess.check_call(f"docker compose -f '{docker_file}' down vault openldap", shell=True)
+        if distutils.spawn.find_executable("docker") and distutils.spawn.find_executable("terraform"):
+            print("Tearing down docker/terraform")
+            docker_file = get_config_file_path("vault-ldap/docker-compose.yml")
+            subprocess.check_call(f"docker compose -f '{docker_file}' down vault openldap", shell=True)
 
-    _test_teardown()
-    _test_setup()
-    request.addfinalizer(_test_teardown)
+    if worker_id == "master":
+        _test_teardown()
+        _test_setup()
+        atexit.register(_test_teardown)
+        return
+
+    root_tmp_dir = tmp_path_factory.getbasetemp().parent
+    fn = root_tmp_dir / "setup"
+    with FileLock(str(fn) + ".lock"):
+        if fn.is_file():
+            return
+        else:
+            if os.path.exists('../../vault.json'): os.unlink('../../vault.json')
+            if os.path.exists('../../vault.json.lock'): os.unlink('../../vault.json.lock')
+            _test_teardown()
+            _test_setup()
+            atexit.register(_test_teardown)
+            fn.write_text(".")
+
